pkgdep/git: Build json-c from nvme-cli's subproject
The nvmecli_plugin ref is based on an older nvme-cli version (2.4),
which does not include several refactors|fixes targeted for the
json print lib. Without these changes, this particular version
of nvme-cli cannot be compiled under distros which don't provide
json-c lib at a expected version (>= 0.13) - this is the case
for centos7 for instance.

To mitigate that, make sure that json-c is built from the nvme-cli's
subproject to provide proper support under less up-to-date distros.

Change-Id: Ida815ba63266c954020bedb7bc7662fc67255f2c
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20379
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>