nvmf/tcp: don't wait until host closes connection
There are several cases where the target should terminate the connection
immediately (e.g. receiving H2CTermReq or host sending mqes+1 requests),
while the only situation in which the target should wait for the host to
do it is after receiving C2HTermReq.

So, a new flag was added that allows for distinguishing between these
two cases.  It ensures that the target actually closes the socket when
necessary, as previously it'd only stop processing new requests, but
would wait until the host terminates the connection.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I8a7ec47f444ef6ad1e69aa20a4a39ecbec48a27c
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20448
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>