accel_perf: use -S for specifying crc32 seed
The help said that the seed can be specified using the -s option, but it
wasn't part of the optstring.  And it cannot be part of the opstring,
because -s is already reserved by the --mem-size option.  So, use -S
instead.

Additionally, this patch renames argc,argv -> ch,arg in parse_args(), as
these variables describe a single command-line parameter, not the whole
vector.

Fixes #3164.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I234377e0eef7f4a4ddcd55a3baeb6dda20b573cb
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20449
Community-CI: Mellanox Build Bot
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>