lib/idxd: DIF strip DSA implementation
Extend the IDXD API with DIF Strip operation support.

The DIF Strip operation copies memory from the source address to
the destination address, removing the Data Integrity Field (DIF).
It optionally computes the DIF on the source data and compares the
computed DIF to the DIF contained in the source data.
DIF computation is performed on each block of source data.

If the operation is completed successfully, the final Reference Tag
and Application Tag is written to the completion record along with
a Success completion status. If an error is detected in the DIF in
the source data, the operation stops.

Change-Id: Ia4e5e08eda565678a38dbc7ad175cb425932a39f
Signed-off-by: Slawomir Ptak <slawomir.ptak@intel.com>
Signed-off-by: Krzysztof Sprzaczkowski <krzysztof.sprzaczkowski@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/22164
Community-CI: Mellanox Build Bot
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>