bdev: add json config saving for bdev_enable_histogram
When the "save_config" RPC command is called in SPDK,
the current environment is saved. However, the environment
created by "bdev_enable_histogram" is not included
in the saved configuration.

In a production environment, it is crucial to monitor
various IO indicators, and IO latency is one of the
important metrics. Fortunately, SPDK provides an RPC
method called "bdev_get_histogram" to obtain the IO
latency histogram of a bdev device. However, before
using this RPC method, it is necessary to first call
the "bdev_enable_histogram" command.

In a production environment, SPDK is typically configured
to automatically restart using the saved configuration
file (by using the "save_config" command in rpc.py)
with the help of systemd or Docker. However, since the
"bdev_enable_histogram" configuration is missing from
the saved configuration, the IO latency monitoring will
fail once SPDK restarts.

Change-Id: I02f217094478190e73ed1ffb8b56793b91050e3d
Signed-off-by: luchangqi.123 <luchangqi.123@bytedance.com>
Signed-off-by: zhenwei pi <pizhenwei@bytedance.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20502
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>