lib/lvol: fix action_in_progress
action_in_progress right now determines if lvol is
in process of being closed or deleted.

It should be set only once blobstore functions are called,
and unset if they do encounter an error afterwards.
Failure to close can only occur during close, then the lvol
remains. For destroy it is freed anyway.

This patch moves setting it to true bit later on close/destroy,
and joins success/error paths by setting in to false in both.

While here remove lvol_free() that should not be freed
in the close path, if blob close failed.
This allows to add relevant UT.

Part of work on #3006.

Change-Id: I9dc8b8defb6090b9e40b3520fbdccbd071622ba6
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18684
Community-CI: Mellanox Build Bot
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>