check_so_deps: suppress types only for single SO_VER
Suppressing types with libabigail is a manual process,
that should be used with care. Otherwise ABI changes
might be missed or ignored.

Main requirement going forward should be to specify
name of the type and specific SO_VER of the library,
when adding new suppression.
Just with this change, the suppress type entry can be
removed at any later time. Rather than keep up with
SO_VER changes.

This leaves a gap during that specific SO_VER.
Where ABI breaking change could be introduced, but would
be ignored due to the suppression.
There are more specific ways to limit the scope of the
suppressions. Most useful ones are has_data_member_*,
where a name of removed field or offset of the field
can be specified. The more specific the better chances
of catching just that change.
Please refer to libabigail manual for [suppress_type].

Three examples in this patch are:
- spdk_app_opts, where reserved185 field name was removed
and new member inserted between 1480-1488 bits
- spdk_nvme_cap_register, where due to limitations of
libabigail it was not possible to refer to change in
bits part of the union and narrow down this suppression
- spdk_bdev_opts, where reserved field name was removed
and new members inserted at 192-224 bits.

As the second example shows, this should be best effort.

Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Change-Id: I8d9bb0fcc4d305426ee1b03d9c43ad20d299ae51
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20544
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>