nvme: do port number checking in nvme_parse_addr()
TCP was already range checking the port number, so
move it to the common nvme_parse_addr() so that RDMA
gets the port checking as well.

Note, previously TCP was checking against MAX_INT,
so change this to reject >= 65536 instead which is
the real IP port limit.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Suggested-by: Alexey Marchuk <alexeymar@nvidia.com>
Change-Id: I91e2f23ac4f4d8ec80cdea95a4bbff73477b8464
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20565
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>