nvme/pcie: drop nvme_qpair_free_request
Just use nvme_free_request instead at the single call site.
This will be optimized further in later patches.

Note that this results in getting the qpair from the req
instead of the tracker, which touches a cacheline
unnecessarily. This will be re-optimized again later in
this patch series.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ic1952c574101d8f2940be7323f3c95daa9e8da80
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20646
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: Michael Haeuptle <michaelhaeuptle@gmail.com>
Reviewed-by: Deepak Abraham Tom <deepak-abraham.tom@hpe.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>