bdev: fix callback not working if unquiesce called inside quiesce cb
A callback for unquiescing a range inside the quiesce callback of the
same range would not be execued. The quiesce_ctx for a range is shared
between quiesce and unquiesce operations and the unquiesce callback, set
by the quiesce cb, was cleared immediately after quiesce cb finished.

Change-Id: I8cc34cb6394e533c536c0f4efb831d8bed8f7bce
Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20699
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>