nvmf: add --ana-state parameter to nvmf_subsystem_add_listener
Currently, when a listener is added to the subsystem, all ANA
(Asymmetric Namespace Access) groups are set to default state
SPDK_NVME_ANA_OPTIMIZED_STATE.
In certain applications, the user may want to submit initial
state for ANA groups during listener creation, and in this case
such a state shall be used as initial one for all ANA groups
created in the listener.

This patch modifies nvmf_subsystem_add_listener RPC call and adds
--ana-state optional parameter to it. If the parameter is present
in RPC nvmf_subsystem_add_listener call, the listener is created
and all 32 ANA groups are initialized to the provided state.
If the parameter is not present, the nvmf_subsystem_add_listener
RPC call works as it did before, and all ANA groups created for
this listener are initialized to SPDK_NVME_ANA_OPTIMIZED_STATE.

The possible values for '--ana-state' (or shortly '-n') are:
--ana-state optimized
--ana-state non_optimized
--ana-state inaccessible

Fixes #3154

Change-Id: I8ce2083698f7dfff44e19d5a5b8a1bcede6b4313
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20663
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Sebastian Brzezinka <sebastian.brzezinka@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot