module/raid: raid_io handling during rebuild
Handle I/Os for the processed (rebuilt) range differently than for the
unprocessed range. Use a copy of raid_bdev_io_channel, corresponding to
the processed range, and assign it to the I/Os for this range. I/Os that
span both ranges will be queued by the bdev layer if the process window
range is quiesced. Otherwise, split them and handle separately.

Change-Id: I802fd5071d37fec07aa43d5dc8ce353aba41a2fe
Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/18741
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>