test/nvmf: test tls only with tcp
Currently we run tls.sh test script in nvmf suite,
regardless of transport type. The tls.sh supports
TCP only, but running it as part of RDMA tests does
not yield failures, because we exit with
'Unsupported transport: rdma' and no error code
(exit 0). This is unintuitive, because we go through
setup phase of `nvmf_tls` test, but never actually
test the functionality.

Adding a check for TCP transport should be enough
to amend this problem.

Change-Id: I86221dfdd40f68979f3385dc518464f021d18bf0
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20773
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>