nvmf: option to add NVM subsystem discovery log entries
It is now possible to add discovery log entries describing NVM
subsystems that aren't available in that NVMe-oF target.

Since this is fairly similar to the way discovery subsystem referrals
are handled, the nvmf_discovery_add_referral is also used for adding
such entries.  Users can now specify an NQN (which defaults to a
discovery subsystem NQN), which is then used to select a subtype.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I608b98e93ae237d118c1a5feecbed713c774b1d8
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/20901
Reviewed-by: Vasuki Manikarnike <vasuki.manikarnike@hpe.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Abhineet Pandey <abhineet.pandey@nutanix.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>