lib/iscsi: add rpc method iscsi_enable_histogram
SPDK provides a histogram function about io delay in the bdev layer,
but when using the iscsi protocol, especially when the io size is large
and io splitting is required, we need to be more concerned about
the end-to-end io latency of the iscsi layer.

An example of WRITE16 command with bulk data:
    I                T
 WRITE16        ->
                <- R2T
 DATAOUT        ->
 DATAOUT        ->
   ...
 DATAOUT(FINAL) ->

Therefore, a new rpc method is added to control the opening or
closing of the histogram function.

Change-Id: Iec96a54a8da2a58c0542e7b38e2161ade505f09f
Signed-off-by: Changqi Lu <luchangqi.123@bytedance.com>
Signed-off-by: zhenwei pi <pizhenwei@bytedance.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21002
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>