test/scheduler: Speed up mapping of the cpu topology
Currently, for systems with multiple nodes and >100 cpu threads
the time needed for map_cpus() to complete is ~4s. The bottleneck
here is the is_cpu_online() which performs quite a lot of steps
to yield such a simple information.

Introduce new wrapper functions that will simply read from cpu's
online attribute with a special case for cpu0 using the old routine.

This drops the time needed to complete the map, on the same system,
from ~4s to ~0.6s

Change-Id: I66e004f5bb7505e47d70c487fe74a1b727856f58
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21030
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Karol Latecki <karol.latecki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>