module/keyring: add keyring_file module
This module allows for using keys stored in files.  Each such file
should only contain a single key and the module copies it without any
modifications.  Additionally, it checks the permissions of the file to
ensure that only the owner of the file is allowed to access it.

This mechanism is very similar to the method employed by the NVMe driver
and the NVMe-oF target to supply NVMe/TLS pre-shared keys.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I6eb15549073e8388b3f806f2888df3b1e3676ad5
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21737
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Seung yeon Shin <syeon.shin@samsung.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>