lib/init: manage multiple RPC servers
RPC library now encapsulates RPC server specific data, so it is
possible to run multiple servers at the same time. To utilize that
in init poll each server that was launched through this library and
tie registering/unregistering of the poller to the existence of at
least one RPC server.

Modifications to existing APIs' behavior:
 * spdk_rpc_initialize() launches a new server instead of
   overwriting existing one,
 * spdk_rpc_finish() stops all existing servers instead of
   stopping a single, global server instance,

Addition of new API:
 * spdk_rpc_server_finish() will stop a server listening
   on a given address (if it exists).

Change-Id: I9c794e31f7ee27eb5607925b6c7a8446387670b7
Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21080
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Community-CI: Mellanox Build Bot