bdev/nvme: use keyring for PSKs
It is now possible to specify NVMe/TLS PSKs via keys attached to the
keyring.  For now, the old method is also available, but it's deprecated
and will be removed in the future.  No new RPC parameters have been
added, instead the PSK is first interpreted as a key name and, if that
fails, as path to the key file.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I663e67ff11a3943c3c11d2f4ba4e31473fcc2e67
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21749
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>