test/iscsi_tgt: Stop looping over $devs, pick single device
In practice, this is how the test progresses under the CI - not only
it explicitly waits for a single device but indeed only one device
becomes available.

Also, this test holds record for the longest test in the pool - on
average it's > 20m. Granted, it spins under nightly only but this
is still a ridiculous amount of time where the bottleneck is
actual SPDK build. Having this test running against multiple devices
would greatly increase the time needed for the job to gracefully
complete (it already hits timeouts of 45m).

Change-Id: I78a43b57507ea62dd63c36da0f6dacd44b9416f6
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21106
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>