CONFIG, pkgdep: enable nvme-cuse by default
nvme-cuse was only gated by a configure flag originally because it was
experimental. It only requires libfuse3, so since it is no longer experimental
this is reasonable to just enable by default.

nvme-cuse is also the only way to use nvme-cli with SPDK, so it is better to
just enable it by default for that reason alone.

Note, we are only enabling CONFIG_NVME_CUSE=y in this patch. CONFIG_FUSE
(for blobfs) remains disabled by default.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ie40f2daf60945e6329e91c8f7dd0aac8e8b7baf2
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21090
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Changpeng Liu <changpeng.liu@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Community-CI: Mellanox Build Bot