test/accel: Add script to test accel RPC methods for DSA module
The accel dsa_scan_accel_module may be performed to use
the DSA Module before starting the framework but double-calling
of this function is not allowed.

The RPC accel_assign_opc method may be performed to override
the operation code assignments to modules before starting
the framework. So, the assigned opcode can be verified after
starting the framework.

Change-Id: I51292520471c3c56a875d80f29d71624ce7b33b7
Signed-off-by: Krzysztof Sprzaczkowski <krzysztof.sprzaczkowski@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21162
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>