autotest: Monitor system resources during functional tests
This is done in order to have a better picture on how target system
behaves during our tests. Hardware-specific tooling will be executed
only under phy nodes, but more generic data will be available for
VMs as well.

Gathered data will include the following:

 - CPU load (VMs and PHY)
 - vmstat (VMs and PHY)
 - Power Management data (PHY only)
 - Thermal data (PHY only)

Note that Power Management and Thermal data may not be available for
all the PHY nodes we have in the CI pool since it solely depends
on specific features provided by BMCs and CPUs (which may not be
supported) but the majority should provide similar data.

Change-Id: Ifa904b4a76b7b2597bee6d44989573c1ed9e06c1
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21174
Reviewed-by: Pawel Piatek <pawelx.piatek@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>