nvme: add new RESET failure reason
We can use this to differentiate between disconnects
due to a reset v. real transport failures.

This is needed because when using nvme multi-process,
a process calling spdk_nvme_ctrlr_reset() cannot
try to reinitialize foreign IO qpairs from other
processes. This patch handles this correctly now.

This required adding a new value to the transport
failure reason enum, which extended the required size
from 2 to 3 bits. So this patch also does some rearranging
to account for the new bits, including shrinking qprio
to 2 bits which is all that is needed based on spec.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I64dfd06c7b0c274ec8f6735bb837eccee658a34d
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21212
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>