test/nvmf: Fix multipath test
This test was improperly parsing the available nvme subsystems
leading to the following:

 target/multipath.sh@25 -- # cat /sys/block/nvme0cn1/ana_state
 cat: /sys/block/nvme0cn1/ana_state: No such file or directory
 ...
 /target/multipath.sh: line 25: [: !=: unary operator expected

In this case  bash was bailing out of the loop condition and
the function was returning successfully (all this because of
improper quoting and use of [).

fio-wrapper was then properly picking up the target devices but the
multipath test was still not verifying their ANA states.

Refactor this test to pick up proper subsystem from the sysfs and
by following this path also its devices.

Change-Id: I614ba52ffef2c67d73c4205356916dba785ab594
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21347
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>