dpdk: When using the DPDK submodule, install it
If the user is using the DPDK submodule, if they run 'make install' then
install DPDK too. Otherwise none of the installed SPDK applications will
work.

For rpmbuild, if the user is building with the DPDK submodule then we
package DPDK libraries into spdk-devel. If they are using an external
DPDK, then we make spdk-devel depend on a suitable DPDK package.

Change-Id: I589e24891e0d13e1c674fae9c76be1df25d9f5aa
Signed-off-by: Ben Walker <ben@nvidia.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21385
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>