test/vhost: Add cpu gen wrapper around conf-generator
The goal here is to not use a static cpu configuration in vhost
tests, but rather determine it dynamically based on the underlying
topology.

Also, change vm[N]_node override var to VM[N]_NODE as existing
implementation didn't really allow the user to override the node
this way (since it was explicitly set by the conf-generator while
constructing initial NUMA-aware setup).

Change-Id: I791afaa8ebbd77de85d1e29c2fa43d442c22cac8
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21420
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>