util: add SPDK_IOV_ONE macro
Having that spdk_iov_one fn is marked deprecated. The issue with that
function is that it enforce iovcnt type which might not be of type int
but u32 or u8 etc. Having that compiler would complain when using the
function with other types than int.

It is a pre-work for layout optmz of nvmf request (iovcnt change there
from int to u8).

Change-Id: Id1534f4d3f60d35c58c26e4f4cf0d605ae11398a
Signed-off-by: Jacek Kalwas <jacek.kalwas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21433
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>