nvmf: don't ignore reservation load errors
Also, change the return value of nvmf_ns_reservation_load_json() in case
the file doesn't exist. It is not a error, as already stated in a
comment in the function, and was only non-zero to skip
nvmf_ns_reservation_restore(). It is safe to call
nvmf_ns_reservation_restore() in this case, because the info struct is
empty so nothing will be restored.

Suggested-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I114f8db97aa5ce9e2c5c87b8b077b2d980464289
Signed-off-by: Artur Paszkiewicz <artur.paszkiewicz@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21489
Community-CI: Mellanox Build Bot
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>