nvmf: add nvmf_ctrlr_get_ns()
Currently _nvmf_subsystem_get_ns() is used to get the spdk_nvmf_ns associated
with a command that is sent to a ctrlr. Namespace masking will require using
the ctrlr to determine if a ns is accessible or not. So prepare for that by
adding nvmf_ctrlr_get_ns() here, which does exactly what
_nvmf_subsystem_get_ns() does today. Then future patches can just change
the implementation of nvmf_ctrlr_get_ns() to add the namespace mask checks.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Iaadffc4fa0619075b3ec5532c2dbcab1331200ec
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21515
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>