nvme/pcie: Force psdt to PRP on admin commands without payload
The NVMe spec mentions:
PRPs shall be used for all Admin commands for NVMe over PCIe
implementations.

The psdt bit may not always be correctly set to PRP. For
example, if a command is received on the nvmf layer and is
forwarded to the nvme layer using passthru APIs, the psdt
bit will be incorrect.

Change-Id: Ib9ae49898230ee619894ca267650abd2503d90f5
Signed-off-by: Alex Michon <amichon@kalrayinc.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21567
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>