lib/util/dif: Reference tag optional check on remap
Add a parameter to skip the reference tag check during the remap
operations.

This is needed in cases, when the check and remap operations happen
in two different places on the I/O path. This is true, for example,
for RAID bdev, where there are multiple levels of LBA remapping happening
on different layers.

Change-Id: Ic7692ddc9337cde93063038e19395ae51b3e537c
Signed-off-by: Slawomir Ptak <slawomir.ptak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/21569
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Community-CI: Mellanox Build Bot
Reviewed-by: Jim Harris <jim.harris@samsung.com>